requirements.txt
fastapi
uvicorn
requests
PyQt6
ddgs
readability-lxml
llama_api_client
pydantic
